functions/export_to_gcs/requirements.txt
pandas
sqlalchemy
pyodbc
google-cloud-storage
